pyproject.toml
[project]
name = "metamed-data-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.26.0",
    "fastapi>=0.100.0",
    "gunicorn>=21.0.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langchain-pinecone>=0.2.12",
    "openai>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pinecone>=7.3.0",
    "psycopg2-binary>=2.9.6",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-json-logger>=2.0.7",
    "python-memcached>=1.62",
    "python-multipart>=0.0.6",
    "requests>=2.28.0",
    "slowapi>=0.1.8",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.23.0",
]
